Extract DuplicateDetector

+++ atto-node/src/main/kotlin/org/atto/node/vote/priotization/VotePrioritizationProperties.kt
@@ -8,5 +8,5 @@ import org.springframework.context.annotation.Configuration
 @ConfigurationProperties(prefix = "atto.vote.prioritization")
 class VotePrioritizationProperties {
     var groupMaxSize: Int? = null
-    var cacheMaxSize: Int? = null
+    var cacheMaxSize: Long? = null
 }
